1. 全局通用布局 2. 路由跳转和高亮同步 3. 全局自定义请求 + 前端代码生成器 4. 清理多余代码

+++ src/main/java/com/morfism/aiappgenerator/common/PageRequest.java
@@ -11,12 +11,12 @@ public class PageRequest {
     /**
      * 当前页号
      */
-    private int pageNum = 1;
+    private long pageNum = 1;
 
     /**
      * 页面大小
      */
-    private int pageSize = 10;
+    private long pageSize = 10;
 
     /**
      * 排序字段
